enhanced DShow filter assemblies for multiple cameras.

+++ src/rgbconv.h
@@ -1,7 +1,7 @@
 #ifndef __RGBCONV_H__
 #define __RGBCONV_H__
 
-unsigned rgb555rgb(void* in, unsigned sz, void** out,unsigned width, unsigned height );
-unsigned rgb565rgb(void* in, unsigned sz, void** out,unsigned width, unsigned height );
+size_t rgb555rgb( void* in, size_t sz, void** out, uint32_t width, uint32_t height );
+size_t rgb565rgb( void* in, size_t sz, void** out, uint32_t width, uint32_t height );
 
 #endif // __RGBCONV_H__
